+++ apps/client/src/app/game/[roomId]/page.tsx
@@ -1,18 +1,14 @@
+'use client';
+
+import { useParams, useSearchParams } from 'next/navigation';
 import React from 'react';
 
 import Game from '@/pages/game/page';
 
-export default async function LobbyPage({
-  params,
-  searchParams,
-}: {
-  params: { roomId: string };
-  searchParams: {
-    host: string;
-  };
-}): Promise<React.JSX.Element> {
-  const host = searchParams.host;
-  const roomId = params.roomId;
+export default function LobbyPage(): React.JSX.Element {
+  const searchParams = useSearchParams();
+  const host = searchParams.get('host') || 'false';
+  const { roomId } = useParams<{ roomId: string }>();
 
   return <Game host={host} roomId={roomId} />;
 }
