+++ apps/client/src/const/keymapping.ts
@@ -1,0 +1,16 @@
+import { KeyMapping } from '@repo/shared-types/src/types';
+
+export const defaultKeyMappings: KeyMapping = {
+  C: 's',
+  'C#': 'e',
+  D: 'd',
+  'D#': 'r',
+  E: 'f',
+  F: 'g',
+  'F#': 'y',
+  G: 'h',
+  'G#': 'u',
+  A: 'j',
+  'A#': 'i',
+  B: 'k',
+};
