+++ apps/client/src/pages/signIn/signin.css
@@ -1,0 +1,129 @@
+@tailwind base;
+@tailwind components;
+@tailwind utilities;
+
+:root {
+  --border-width: 1px;
+  --border-radius: 0.5rem;
+  --color-error: #c94b4b;
+  --color-info: #157efb;
+  --color-info-hover: #0f6ddb;
+  --color-info-text: #fff;
+}
+
+.__next-auth-theme-auto,
+.__next-auth-theme-light {
+  --color-background: #ececec;
+  --color-background-hover: hsla(0, 0%, 93%, 0.8);
+  --color-background-card: #fff;
+  --color-text: #000;
+  --color-primary: #444;
+  --color-control-border: #bbb;
+  --color-button-active-background: #f9f9f9;
+  --color-button-active-border: #aaa;
+  --color-separator: #ccc;
+}
+
+.__next-auth-theme-dark {
+  --color-background: #161b22;
+  --color-background-hover: rgba(22, 27, 34, 0.8);
+  --color-background-card: #0d1117;
+  --color-text: #fff;
+  --color-primary: #ccc;
+  --color-control-border: #555;
+  --color-button-active-background: #060606;
+  --color-button-active-border: #666;
+  --color-separator: #444;
+}
+
+@media (prefers-color-scheme: dark) {
+  .__next-auth-theme-auto {
+    --color-background: #161b22;
+    --color-background-hover: rgba(22, 27, 34, 0.8);
+    --color-background-card: #0d1117;
+    --color-text: #fff;
+    --color-primary: #ccc;
+    --color-control-border: #555;
+    --color-button-active-background: #060606;
+    --color-button-active-border: #666;
+    --color-separator: #444;
+  }
+}
+
+a.button,
+button {
+  background-color: var(--provider-dark-bg, var(--color-background));
+  color: var(--provider-dark-color, var(--color-primary));
+}
+
+a.button:hover,
+button:hover {
+  background-color: var(
+    --provider-dark-bg-hover,
+    var(--color-background-hover)
+  ) !important;
+}
+
+body {
+  background-color: var(--color-background);
+  font-family:
+    ui-sans-serif,
+    system-ui,
+    -apple-system,
+    BlinkMacSystemFont,
+    Segoe UI,
+    Roboto,
+    Helvetica Neue,
+    Arial,
+    Noto Sans,
+    sans-serif;
+  margin: 0;
+  padding: 0;
+}
+
+h1,
+p {
+  color: var(--color-text);
+  margin-bottom: 1.5rem;
+  padding: 0 1rem;
+}
+
+input[type],
+label {
+  color: var(--color-text);
+}
+
+input[type] {
+  background: var(--color-background-card);
+  border: var(--border-width) solid var(--color-control-border);
+  border-radius: var(--border-radius);
+  font-size: 1rem;
+  padding: 0.5rem 1rem;
+  width: 100%;
+}
+
+input[type]:focus {
+  box-shadow: none;
+}
+
+button span {
+  flex-grow: 1;
+}
+
+.page {
+  display: grid;
+  height: 100%;
+  place-items: center;
+  position: absolute;
+  width: 100%;
+}
+
+.card {
+  background-color: var(--color-background-card);
+  border-radius: 2rem;
+  padding: 1.25rem 2rem;
+}
+
+.provider {
+  margin-top: 1rem;
+}
