+++ package.json
@@ -44,7 +44,6 @@
       "serve-static@<1.16.0": ">=1.16.0",
       "express@<4.20.0": ">=4.20.0",
       "path-to-regexp@<0.1.10": ">=0.1.10",
-      "next@>=14.0.0 <14.2.10": ">=14.2.10",
       "cookie@<0.7.0": ">=0.7.0"
     }
   }
